llm-vs-llm/requirements.txt
transformers>=4.41.0
accelerate>=0.30.0
datasets>=2.19.0
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
openai>=1.0.0